+++ packages/uniforms-bridge-graphql/src/GraphQLBridge.ts
@@ -29,12 +29,6 @@ export default class GraphQLBridge extends Bridge {
     this.validator = validator;
   }
 
-  // This bridge has 3 arguments, so it cannot be constructed implicite in the
-  // createSchemaBridge.
-  static check(/* schema */) {
-    return false;
-  }
-
   getError(name, error) {
     return (
       (error &&