+++ packages/uniforms-bridge-graphql/src/register.ts
@@ -1,6 +1,0 @@
-import { createSchemaBridge } from 'uniforms';
-
-import GraphQLBridge from './GraphQLBridge';
-
-// Register bridge.
-createSchemaBridge.register(GraphQLBridge);