+++ packages/uniforms-bridge-json-schema/src/index.ts
@@ -1,2 +1,1 @@
-import './register';
 export { default, default as JSONSchemaBridge } from './JSONSchemaBridge';