+++ packages/uniforms-bridge-json-schema/src/register.ts
@@ -1,6 +1,0 @@
-import { createSchemaBridge } from 'uniforms';
-
-import JSONSchemaBridge from './JSONSchemaBridge';
-
-// Register bridge.
-createSchemaBridge.register(JSONSchemaBridge);