+++ packages/uniforms-bridge-simple-schema/src/SimpleSchemaBridge.ts
@@ -13,17 +13,6 @@ export default class SimpleSchemaBridge extends Bridge {
     this.schema = schema;
   }
 
-  static check(schema) {
-    return (
-      schema &&
-      schema.getDefinition &&
-      schema.messageForError &&
-      schema.objectKeys &&
-      schema.validator &&
-      schema.version !== 2
-    );
-  }
-
   getError(name, error) {
     return (
       (error &&